+++ src/features/optional/graphics/holidayHats.ts
@@ -5,6 +5,8 @@ import { Holiday } from "../../../enums/Holiday";
 import { config } from "../../../modConfigMenu";
 import { HOLIDAY_TO_NULL_ITEM_ID } from "../../../objects/holidayToNullItemID";
 
+// - Christmas is: Sunday, December 25th
+// - Two weeks before is: Sunday, December 11th
 const CURRENT_HOLIDAY = Holiday.NONE;
 
 // ModCallback.POST_GAME_STARTED (15)